+++ src/sqlfluff/parser_2/lexer.py
@@ -147,9 +147,9 @@ class Lexer(object):
             RegexMatcher.from_shorthand("whitespace", r"[\t ]*"),
             RegexMatcher.from_shorthand("inline_comment", r"(-- |#)[^\n]*"),
             RegexMatcher.from_shorthand("block_comment", r"/\*([^\\]|\\[^\*])*"),
-            RegexMatcher.from_shorthand("single_quote", r"'[^']*'"),
-            RegexMatcher.from_shorthand("double_quote", r'"[^"]*"'),
-            RegexMatcher.from_shorthand("back_quote", r"`[^`]*`"),
+            RegexMatcher.from_shorthand("single_quote", r"'[^']*'", is_code=True),
+            RegexMatcher.from_shorthand("double_quote", r'"[^"]*"', is_code=True),
+            RegexMatcher.from_shorthand("back_quote", r"`[^`]*`", is_code=True),
             SingletonMatcher.from_shorthand("newline", "\n"),
             SingletonMatcher.from_shorthand("dot", ".", is_code=True),
             SingletonMatcher.from_shorthand("comma", ",", is_code=True),
