+++ test/parser_2_segments_file_test.py
@@ -95,7 +95,7 @@ def test__parser_2__base_file_parse(raw, caplog):
             )),)),))
         ),
         (
-            'select a,b, c from sch.blah',
+            'select a,b, c from sch."blah"',
             ('file', (('statement', (('select_statement', (
                 ('keyword', 'select'),
                 ('select_target_group', (
@@ -111,7 +111,7 @@ def test__parser_2__base_file_parse(raw, caplog):
                         ('identifier', (
                             ('naked_identifier', 'sch'),
                             ('dot', '.'),
-                            ('naked_identifier', 'blah'),
+                            ('quoted_identifier', '"blah"'),
                         )),
                     ))
                 )),
